requirements.txt
Babel==2.6.0
Django==2.2
django-macaddress==1.5.0
django-phonenumber-field==2.3.1
django-recaptcha==2.0.2
mod-wsgi==4.6.5
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
netaddr==0.7.19
phonenumbers==8.10.8
protobuf==3.7.1
pytz==2018.9
schedule==0.6.0
six==1.12.0
sqlparse==0.3.0